+++ Spect.Net.SpectrumEmu.Test/Tape/TzxPlayerHelper.cs
@@ -79,13 +79,13 @@ namespace Spect.Net.SpectrumEmu.Test.Tape
             const int HEADER_PILOT_COUNT = TzxStandardSpeedDataBlock.HEADER_PILOT_COUNT;
             const int SYNC_1_PL = TzxStandardSpeedDataBlock.SYNC_1_PL;
             const int SYNC_2_PL = TzxStandardSpeedDataBlock.SYNC_2_PL;
-            const ulong PILOT_END = PILOT_PL * HEADER_PILOT_COUNT;
+            const long PILOT_END = PILOT_PL * HEADER_PILOT_COUNT;
             const int TERM_SYNC = TzxStandardSpeedDataBlock.TERM_SYNC;
             const int PAUSE_MS = TzxStandardSpeedDataBlock.PAUSE_MS;
 
             var start = block.StartTact;
             // --- Skip all pilot pulses + the first sync pulse
-            for (ulong pos = 0; pos < PILOT_END + SYNC_1_PL; pos += 50)
+            for (long pos = 0; pos < PILOT_END + SYNC_1_PL; pos += 50)
             {
                 block.GetEarBit(start + pos);
             }
@@ -111,7 +111,7 @@ namespace Spect.Net.SpectrumEmu.Test.Tape
 
             // --- Play back the pause
             nextTact = block.LastTact;
-            for (var pos = nextTact; pos < nextTact + (ulong)PAUSE_MS * block.PauseAfter + 100; pos += 50)
+            for (var pos = nextTact; pos < nextTact + PAUSE_MS * block.PauseAfter + 100; pos += 50)
             {
                 block.GetEarBit(pos);
             }
@@ -122,19 +122,19 @@ namespace Spect.Net.SpectrumEmu.Test.Tape
         /// </summary>
         /// <param name="block">Data block to play back</param>
         /// <remarks>Last tact position</remarks>
-        public static ulong ReadUntilPause(this TzxStandardSpeedDataBlock block)
+        public static long ReadUntilPause(this TzxStandardSpeedDataBlock block)
         {
             const int PILOT_PL = TzxStandardSpeedDataBlock.PILOT_PL;
             const int HEADER_PILOT_COUNT = TzxStandardSpeedDataBlock.HEADER_PILOT_COUNT;
             const int SYNC_1_PL = TzxStandardSpeedDataBlock.SYNC_1_PL;
             const int SYNC_2_PL = TzxStandardSpeedDataBlock.SYNC_2_PL;
-            const ulong PILOT_END = PILOT_PL * HEADER_PILOT_COUNT;
+            const long PILOT_END = PILOT_PL * HEADER_PILOT_COUNT;
             const int TERM_SYNC = TzxStandardSpeedDataBlock.TERM_SYNC;
             const int PAUSE_MS = TzxStandardSpeedDataBlock.PAUSE_MS;
 
             var start = block.StartTact;
             // --- Skip all pilot pulses + the first sync pulse
-            for (ulong pos = 0; pos < PILOT_END + SYNC_1_PL; pos += 50)
+            for (long pos = 0; pos < PILOT_END + SYNC_1_PL; pos += 50)
             {
                 block.GetEarBit(start + pos);
             }
@@ -159,7 +159,7 @@ namespace Spect.Net.SpectrumEmu.Test.Tape
             }
 
             // --- Play back the pause
-            var lastPos = block.LastTact + (ulong)PAUSE_MS * block.PauseAfter + 100;
+            var lastPos = block.LastTact + PAUSE_MS * block.PauseAfter + 100;
             block.GetEarBit(lastPos);
             return lastPos;
         }
